# do not reset camera unless plotter has been shown