… tests/plotting/image_cache/set_focus.png