## Overview

In `add_mesh` we recalculate the camera bounds each and every time we add a mesh. This greatly increases the amount of time it takes to add many meshes and slows down the plotting significantly.

This PR proposes a simple fix to correct this behavior by moving out the reset cliping range to `reset_camera`.

It also includes a minor performance improvement by not checking if an actor exists when adding a new actor with `add_mesh`. Actors are guaranteed to be new actors as we always create a new mesh, even if it may only be a shallow one.

## Timings and code

```py
import pyvista as pv


cubes = [pv.Cube(center=(i, i, i)) for i in range(1000)]

def many_cubes():
    pl = pv.Plotter(off_screen=True)
    for cube in cubes:
        pl.add_mesh(cube)
    pl.show()
```

### This Branch

```
>>> timeit many_cubes()
501 ms ± 2.83 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### `main`

```
>>> timeit many_cubes()
1.45 s ± 24.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```