@MatthewFlamm 2022-07-18
I haven't read the code changes in detail yet, so reacting mostly to the description.  It sounds like the camera will no longer reset everytime a mesh is added, but it will be reset when `plotter.show()` is called.  This behavior should be added to the docstring for arg `reset_camera`, where now it is unclear what the current behavior is.

In case we prefer the current behavior for adding single meshes one at a time, we could accept a list of meshes. For a list of meshes, only update the camera after the last mesh is added.  This might also open up the possibility of getting further efficiencies when using the same set of args for each mesh.