@akaszynski 2022-07-19
To dos:
- Update `reset_camera` docs
- Use do not remove existing actor if name is unset in `add_mesh`