@banesullivan 2022-07-22
This is a substantial performance increase - good catch